// nuxtui/utils/client/types.gen.ts
export type WeekTotals = {
    month: string;
    hours: number;
    milkings: number;
};

export type WorkingHoursRequest = {
    id?: number | null;
    date: string;
    hours: number;
    milkings?: number | null;
    description: string;
    submitted?: boolean | null;
};

export type WorkingHoursResponse = {
    id: number;
    date: string;
    hours: number;
    milkings: number;
    description: string;
    submitted: boolean;
};

export type WorkingHoursWeekOverviewResponse = {
    year: number;
    week: number;
    week_start: string;
    week_end: string;
    working_hours: Array<WorkingHoursResponse>;
    sum_hours: number;
    sum_milkings: number;
    submitted: boolean;
};